script/requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
Faker==37.3.0
idna==3.10
mailslurp-client==16.1.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
tzdata==2025.2
urllib3==2.4.0
